+++ docs/src/global/app.ts
@@ -30,3 +30,4 @@ import '@deckdeckgo/drag-resize-rotate';
 import '@deckdeckgo/demo';
 import '@deckdeckgo/math';
 import '@deckdeckgo/word-cloud';
+import '@deckdeckgo/markdown';