+++ remote/src/global/app.ts
@@ -12,6 +12,7 @@ import '@deckdeckgo/drag-resize-rotate';
 import '@deckdeckgo/demo';
 import '@deckdeckgo/math';
 import '@deckdeckgo/word-cloud';
+import '@deckdeckgo/markdown';
 
 import '@deckdeckgo/slide-title';
 import '@deckdeckgo/slide-author';