+++ studio/src/app/handlers/editor/events/deck/deck-events.handler.tsx
@@ -19,7 +19,6 @@ import {Utils} from '../../../../utils/core/utils';
 import {Resources} from '../../../../utils/core/resources';
 
 import {SlotUtils} from '../../../../utils/editor/slot.utils';
-import {SlotType} from '../../../../utils/editor/slot-type';
 import {ParseElementsUtils} from '../../../../utils/editor/parse-elements.utils';
 
 import {DeckService} from '../../../../services/data/deck/deck.service';
@@ -61,6 +60,7 @@ export class DeckEventsHandler {
       this.el.addEventListener('slideDelete', this.onSlideDelete, false);
       this.el.addEventListener('codeDidChange', this.onCustomEventChange, false);
       this.el.addEventListener('mathDidChange', this.onCustomEventChange, false);
+      this.el.addEventListener('markdownDidChange', this.onCustomEventChange, false);
       this.el.addEventListener('wordCloudDidChange', this.onCustomEventChange, false);
       this.el.addEventListener('imgDidChange', this.onCustomEventChange, false);
       this.el.addEventListener('linkCreated', this.onCustomEventChange, false);
@@ -85,6 +85,7 @@ export class DeckEventsHandler {
     this.el.removeEventListener('slideDelete', this.onSlideDelete, true);
     this.el.removeEventListener('codeDidChange', this.onCustomEventChange, true);
     this.el.removeEventListener('mathDidChange', this.onCustomEventChange, true);
+    this.el.removeEventListener('markdownDidChange', this.onCustomEventChange, true);
     this.el.removeEventListener('wordCloudDidChange', this.onCustomEventChange, true);
     this.el.removeEventListener('imgDidChange', this.onCustomEventChange, true);
     this.el.removeEventListener('linkCreated', this.onCustomEventChange, true);
@@ -1013,7 +1014,7 @@ export class DeckEventsHandler {
 
     elements.forEach((e: HTMLElement) => {
       if (e.nodeName && e.nodeType === 1 && e.hasAttribute('slot')) {
-        if (e.nodeName.toLowerCase() === SlotType.CODE || e.nodeName.toLowerCase() === SlotType.MATH || e.nodeName.toLowerCase() === SlotType.WORD_CLOUD) {
+        if (SlotUtils.isNodeEditable(e)) {
           e.setAttribute('editable', attrEditableValue);
         } else if (ParseElementsUtils.isElementContentEditable(e)) {
           e.setAttribute('contentEditable', attrEditableValue);