+++ studio/src/app/utils/editor/slot.utils.tsx
@@ -35,11 +35,24 @@ export class SlotUtils {
     return selectedElement?.nodeName?.toLowerCase() === SlotType.WORD_CLOUD;
   }
 
+  static isNodeMarkdown(selectedElement: HTMLElement | null): boolean {
+    return selectedElement?.nodeName?.toLowerCase() === SlotType.MARKDOWN;
+  }
+
   static isSlotTypeEditable(type: SlotType): boolean {
     return type !== SlotType.IMG && type !== SlotType.SOCIAL && type !== SlotType.DEMO;
   }
 
   static isNodeDragDropResize(selectedElement: HTMLElement | null): boolean {
     return selectedElement?.nodeName?.toLowerCase() === SlotType.DRAG_RESIZE_ROTATE;
+  }
+
+  static isNodeEditable(element: HTMLElement): boolean {
+    return (
+      element.nodeName.toLowerCase() === SlotType.CODE ||
+      element.nodeName.toLowerCase() === SlotType.MATH ||
+      element.nodeName.toLowerCase() === SlotType.WORD_CLOUD ||
+      element.nodeName.toLowerCase() === SlotType.MARKDOWN
+    );
   }
 }