+++ webcomponents/markdown/stencil.config.ts
@@ -1,19 +1,18 @@
-import { Config } from '@stencil/core';
+import {Config} from '@stencil/core';
 
 import {sass} from '@stencil/sass';
 import {postcss} from '@stencil/postcss';
 import autoprefixer from 'autoprefixer';
 
 export const config: Config = {
-  namespace: 'markdown-parser',
+  namespace: 'deckdeckgo-markdown',
   outputTargets: [
     {
       type: 'dist',
-      esmLoaderPath: '../loader',
     },
     {
       type: 'www',
-      serviceWorker: null, // disable service workers
+      serviceWorker: null,
     },
   ],
   plugins: [